+++ src/wfpt.pyx
@@ -159,8 +159,8 @@ def wiener_like_rlddm(np.ndarray[double, ndim=1] x,
 # JY added on 2021-12-05 for two-step tasks
 def wiener_like_rlddm_2step(np.ndarray[double, ndim=1] x1, # 1st-stage RT
                       np.ndarray[double, ndim=1] x2, # 2nd-stage RT
-                      np.ndarray[double,ndim=1] s1, # 1st-stage state
-                      np.ndarray[double,ndim=1] s2, # 2nd-stage state
+                      np.ndarray[long,ndim=1] s1, # 1st-stage state
+                      np.ndarray[long,ndim=1] s2, # 2nd-stage state
                       np.ndarray[long, ndim=1] response1,
                       np.ndarray[long, ndim=1] response2,
                       np.ndarray[double, ndim=1] feedback,
